+++ components/ble_host/include/ble_types/types.h
@@ -19,12 +19,12 @@ extern "C" {
 typedef signed char      s8_t;
 typedef signed short     s16_t;
 typedef signed int       s32_t;
-typedef signed long long s64_t;
+typedef int64_t          s64_t;
 
 typedef unsigned char      u8_t;
 typedef unsigned short     u16_t;
 typedef unsigned int       u32_t;
-typedef unsigned long long u64_t;
+typedef uint64_t           u64_t;
 
 #ifndef _SSIZE_T_DECLARED
 typedef long ssize_t;
